Struct coatcheck::Ticket

#[must_use = "you need this ticket to claim your item"]
pub struct Ticket {
    // some fields omitted
}

A Ticket is an opaque data structure that can be used to claim the associated value.

Note: Tickets can't be copied to prevent re-use (a ticket can only be exchanged for exactly one item).